Improve code readability in time library
